refactor(render): make BuiltinParticle shared & format files with IDEA
